Don't use passive movement to train
Passive movement is not the best way to train it, since the spatial and temporal characteristics of the resulting brain signal are different between performed/imagined movement and the sensory/proprioceptive activations from passive movement. Several signals like premotor activations and the Bereitschaftspotential are not present. Subjects that cannot move themselves, better calibrate using imagined movements they intend themselves, whether or not these are supported by an external actuator.